Split lexer into a separate file
