+++ source/CMakeLists.txt
@@ -1,6 +1,8 @@
 find_package(nlohmann_json 3.2.0 REQUIRED)
 
 add_executable(sapc
+    lexer.cc
+    lexer.hh
     location.cc
     location.hh
     main.cc
